Improved All Dungeons and Renamed SK-XXX
Fixed the bounding Eternal Fabric that had been clipped off in some
dungeons due to an export bug in Dimensional Doors. Re-exported other
dungeons that did not have bounds marked by Eternal Fabric so that the
schematic dimensions would be minimal - just large enough to enclose the
schematics. This related to another DD issue that caused dungeons
without permafabric bounds to default to very large sizes. It's
important to use actual sizes since I'm hoping to restrict dungeon
generation in the future to prevent dungeon clipping around the top and
bottom of the map. Finally, renamed SK-XXX to SK-Claustrophobia.