Added SK-AnchoredDescent and SK-EyesOfTricksters
Added two new dungeons. SK-AnchoredDescent appears to be an exit room
but it actually has a hidden dimensional door off the obvious path.

SK-EyesOfTricksters is a simple puzzle room. I experimented with using
tripwires as signal carriers (as opposed to redstone dust). They're much
harder to see and they don't have to be set on top of solid blocks.
However, they need something to hit the wire in order to activate. I
dropped gravel to do that. In order to stretch out the activation time,
I placed cobwebs under the wires to slow the gravel's fall. The dungeon
also features a new design for tamper-detecting pistons. Previously, I
used repeaters to create clocks, but in this case, I used two hoppers
looping an item back and forth and a comparator beside one of them. This
creates a compact clock to drive the piston's polling. It also seems
that this mechanism doesn't need to be "started up" with an external
update like repeater-based designs.