Added SK-NicolesTower
Added the new dungeon SK-NicolesTower. This was an old idea - a tower
made mostly of stone bricks and silverfish blocks. The outside of the
tower has a spiral staircase and pistons randomly push the sides of the
tower outward. I named it after Nicole, Java's girlfriend, who was
spectating as a I built it.